refactoring app.js
