+++ client/index.html
@@ -116,5 +116,6 @@
 <!-- <script async type="text/javascript" src="src/wasm/function.js"></script> -->
 <script async type="text/javascript" src="src/wasm/geometry.js"></script>
 <script src="src/app.js" type="module" defer></script>
+<script src="src/autocomplete.js" defer></script>
 
 </html>
