+++ client/src/config.js
@@ -1,0 +1,9 @@
+export const firebaseConfig = {
+    apiKey: 'AIzaSyA7oxGdbv2i-fqEdNYG2D_fCpyfP96-uE0',
+    authDomain: 'poflow-leta-tech.firebaseapp.com',
+    projectId: 'poflow-leta-tech',
+    storageBucket: 'poflow-leta-tech.appspot.com',
+    messagingSenderId: '545726653232',
+    appId: '1:545726653232:web:b2b8a27aba7f00235c9ef5',
+    measurementId: 'G-GSENSHJ9BM'
+};
