+++ client/src/poflow.js
@@ -1,0 +1,5 @@
+export const GetEllipseCoordinates =
+    Module.cwrap('GetEllipseCoordinates', 'string', ['number', 'number']);
+
+export const GetNaca00XXCoordinates =
+    Module.cwrap('GetNaca00XXCoordinates', 'string', ['number', 'number']);
